MyLead

+++ database/migrations/2020_12_20_113723_create_prices_table.php
@@ -15,7 +15,7 @@ class CreatePricesTable extends Migration
     {
         Schema::create('prices', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('price')->unique();
+            $table->integer('price')->unique();
             $table->timestamps();
         });
     }
